fix(ui): hoist popstate listener into useBrowserHistory so back/forward works in every view
- the UI's only popstate listener lived in MainApp, which App() never
  renders for the terms/telegram-bind/slack-bind views; Back from
  Settings -> "View Terms of Use" left the URL at /settings with the
  Terms page still rendered and no way out but a reload
- useBrowserHistory() is called from App() above the early returns, so
  the listener survives every view
- chat teardown is transition-based: leaving chat for any view (not
  just list) resets chat context and clears selectedAgent
- TermsView's Back button becomes an SPA navigation; its Accept handler
  keeps the deliberate full reload that clears the 412 terms_stale gate

Refs #1219
