+++ packages/ui/src/app.tsx
@@ -11,10 +11,7 @@ import { ArtifactsView } from "./modules/artifacts/views/artifacts-view.js";
 import { ExperimentsListView } from "./modules/experiments/views/experiments-list-view.js";
 import { KnowledgeBaseConfigView } from "./modules/knowledge-bases/views/knowledge-base-config-view.js";
 import { KnowledgeBasesListView } from "./modules/knowledge-bases/views/knowledge-bases-list-view.js";
-import {
-  parseRoute,
-  routeToNavigationState,
-} from "./modules/platform/lib/routes.js";
+import { useBrowserHistory } from "./modules/platform/hooks/use-browser-history.js";
 import { useFirstRunRedirect } from "./modules/sandboxes/hooks/use-first-run-redirect.js";
 import { SandboxHomeView } from "./modules/sandboxes/views/sandbox-home-view.js";
 import { SandboxWizardView } from "./modules/sandboxes/views/sandbox-wizard-view.js";
@@ -28,6 +25,10 @@ import { useStore } from "./store.js";
 export default function App() {
   const view = useStore((s) => s.view);
   const theme = useStore((s) => s.theme);
+
+  // Must stay above the early returns: the terms/bind views render without
+  // MainApp, and back/forward has to keep working there too.
+  useBrowserHistory();
 
   // Apply theme on mount + listen for system preference changes
   useEffect(() => {
@@ -79,36 +80,6 @@ function MainApp() {
         message: "Connection authorized.",
       });
     }
-  }, []);
-
-  // Browser back/forward
-  useEffect(() => {
-    const enterChat = (agentId: string) => {
-      useStore.getState().resetChatContext();
-      useStore.setState({ selectedAgent: agentId, view: "chat" });
-    };
-    const leaveChat = () => {
-      useStore.getState().resetChatContext();
-      useStore.setState({ selectedAgent: null, view: "list" });
-    };
-    const onPopState = () => {
-      const route = parseRoute(window.location.pathname);
-      if (route.view === "chat") return enterChat(route.agent);
-      if (route.view === "knowledge-base-chat") {
-        useStore.getState().resetChatContext();
-        useStore.setState({
-          selectedAgent: route.agent,
-          view: "knowledge-base-chat",
-        });
-        return;
-      }
-      // Unknown paths resolve to "list"; leaveChat also tears down chat context.
-      if (route.view === "list") return leaveChat();
-      useStore.setState(routeToNavigationState(route));
-    };
-    onPopState();
-    window.addEventListener("popstate", onPopState);
-    return () => window.removeEventListener("popstate", onPopState);
   }, []);
 
   // Chat owns its mobile sessions/chat nav, so the rail hides its bottom bar
